backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
pydantic-settings
httpx
python-jose[cryptography]
passlib
google-auth
